# services/quiz/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API квизов, вопросов, ответов и пользователей quiz-service.\nПубличные ответы не содержат is_correct, а пользователи — хешей паролей.",
        "title": "QuizGo CRUD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "JPG/PNG до 5 MiB, до 20 мегапикселей. Поле multipart: file. Возвращает ID и временную ссылку S3.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Загрузить изображение вопроса",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quiz.UploadedImage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Перенаправление на временную ссылку S3",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "Ссылка S3"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "images"
                ],
                "summary": "Удалить неиспользуемое изображение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{id}/url": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получить свежую ссылку на изображение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.UploadedImage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/quizzes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Мои квизы с количеством вопросов и датой изменения",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quiz.QuizSummary"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quizzes/{id}": {
            "get": {
                "description": "Возвращает квиз с вопросами и ответами без признака is_correct.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Получить квиз",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.QuizResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID квиза",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Квиз не найден",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Удалить квиз",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quizzes/{id}/author": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Содержимое квиза для автора, включая правильные ответы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.QuizContent"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quizzes/{id}/content": {
            "get": {
                "description": "Порядок массивов соответствует порядку вопросов и ответов. is_correct отсутствует. image_url — временная подписанная ссылка S3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Содержимое квиза для просмотра",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quizhandler.PublicContent"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Требуется актуальная revision, иначе 409. ID=0 создаёт элемент; отсутствующие в массивах элементы удаляются. Порядок задаётся массивами. Пустой квиз допустим. До 100 вопросов, в каждом 2–8 ответов и ровно один правильный. Название до 50 символов, тексты до 200. time_limit: 5,10,15,20,30,45,60,90,120 секунд.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quizzes"
                ],
                "summary": "Сохранить квиз целиком",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID квиза",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Содержимое и версия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quiz.SaveQuizContent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.QuizContent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "quiz.QuizContent": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "question_count": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz.ContentQuestion"
                    }
                },
                "revision": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "quiz.QuizSummary": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "question_count": {
                    "type": "integer"
                },
                "revision": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "quiz.SaveQuizContent": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz.ContentQuestion"
                    }
                },
                "revision": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "quiz.UploadedImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "quizhandler.PublicAnswerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "question_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Париж"
                }
            }
        },
        "quizhandler.PublicContent": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "question_count": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quizhandler.PublicContentQuestion"
                    }
                },
                "revision": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "quizhandler.QuestionResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "description": "null, если ответы отсутствуют или не были загружены.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quizhandler.PublicAnswerResponse"
                    },
                    "x-nullable": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_id": {
                    "type": "string",
                    "example": "image-1"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "quiz_id": {
                    "type": "integer",
                    "example": 1
                },
                "text_content": {
                    "type": "string",
                    "example": "Столица Франции?"
                }
            }
        },
        "quizhandler.QuizResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 42
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quizhandler.QuestionResponse"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "География"
                }
            }
        },
        "quizhandler.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                },
                "username": {
                    "type": "string",
                    "example": "alice"
                }
            }
        },
        "respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите Bearer \u003cJWT\u003e. Токен можно получить через POST /auth/login с JSON {\"username\":\"alice\",\"password\":\"secret123\"}. Поле token уже содержит префикс Bearer. Регистрация: POST /auth/register с теми же полями.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}